+++ internal/service/order.go
@@ -26,3 +26,13 @@ func (s Service) UserOrderReservation(ctx context.Context, req *spec.UserOrderRe
 	}
 	return &spec.UserOrderReservationResponse{Success: true}, nil
 }
+
+func (s Service) UserOrderRevenue(ctx context.Context, req *spec.UserOrderRevenueRequest) (*spec.UserOrderRevenueResponse, error) {
+	userReservation := entity.UserRevenue{ID: req.Id, ServiceID: req.ServiceId, OrderID: req.OrderId, Cost: req.Cost}
+	err := s.useCase.UserOrderRevenue(ctx, &userReservation)
+	if err != nil {
+		s.log.Errorf("service - Service - UserOrderRevenue - s.useCase.UserOrderRevenue %v", err)
+		return &spec.UserOrderRevenueResponse{Success: false}, errors.Unwrap(err)
+	}
+	return &spec.UserOrderRevenueResponse{Success: true}, nil
+}
