+++ internal/usecase/usecase.go
@@ -9,26 +9,32 @@ import (
 )
 
 type UseCase struct {
-	repo      postgres.Repository
-	txService *TransactionServiceImpl
-	log       *logrus.Logger
+	User
+	Order
+	Report
 }
 
-type Usecase interface {
-	UserBalanceAccrual(ctx context.Context, userDTO *entity.User) error
+type User interface {
 	CreateUser(ctx context.Context, userDTO *entity.User) error
 	GetBalance(ctx context.Context, dto *entity.User) error
-	UserOrderRevenue(ctx context.Context, revenue *entity.UserRevenue) error
-	UserOrderDeleteReservation(ctx context.Context, reservation *entity.UserReservation) error
+	UserBalanceAccrual(ctx context.Context, userDTO *entity.User) error
+}
+
+type Order interface {
 	UserOrderReservation(ctx context.Context, reservation *entity.UserReservation) error
+	UserOrderDeleteReservation(ctx context.Context, reservation *entity.UserReservation) error
+	UserOrderRevenue(ctx context.Context, revenue *entity.UserRevenue) error
+}
+
+type Report interface {
 	CreateMonthReport(ctx context.Context, year uint32, month uint32) error
 }
 
 func New(repo *postgres.Repo, log *logrus.Logger, postgresDB *sqlx.DB) *UseCase {
 	txService := NewTransactionService(postgresDB, log)
 	return &UseCase{
-		repo:      repo,
-		log:       log,
-		txService: txService,
+		User:   NewUserUseCase(repo, log, txService),
+		Order:  NewOrderUseCase(repo, log, txService),
+		Report: NewReportUseCase(repo, log, txService),
 	}
 }
